limepcie: fix building with CMake older v3.20
CMake has different behavior of the `target_source` directive across
versions. In particular, in older versions (v3.19 and older) the
directive is supported only for targets previously declared via
`add_executable` or `add_library`. But starting from v3.20 it also
supports targets declared via `add_custom_target`.

See for reference:
    https://cmake.org/cmake/help/v3.20/command/target_sources.html

As a result, when the module is built on hosts with old CMake it leads
to the following configuration errors:
```
    target_sources called with non-compatible target type
```
and the limepcie module can not be built.

To fix the issue it is proposed to specify the sources for the kernel
module target explicitly via the `set_property` directive.

Test: - try to build the module on a host with old CMake (Debian 11,
        CMake v3.18.4) with the following command:
        $ cmake -DCMAKE_BUILD_TYPE=Release \
                -DBUILD_GUI=OFF \
                -DBUILD_CLI=OFF \
                -DBUILD_CORE=OFF \
                -DBUILD_DOCS=OFF \
                -DBUILD_DRIVERS=ON \
                -DBUILD_PLUGINS=OFF \
                -DINSTALL_UDEV_RULES=OFF \
                -DENABLE_UNIT_TESTS=OFF \
                -DLIMEPCIE_INSTALL=OFF \
                 ../LimeSuiteNG
      - try to build the module on a host with modern CMake (Ubuntu
        24.04, CMake v3.28.3) using the above command;
      - ensure that binaries (limepcie.ko) built with and without the
        patch are the same:
        $ md5sum drivers/linux/limepcie/limepcie-0.1.8/limepcie.ko
        $ 804c702e1bbf40a544cdc5cc30d092ab \
          drivers/linux/limepcie/limepcie-0.1.8/limepcie.ko
Signed-off-by: Gluttton <gluttton@ukr.net>