Fix GUI compilation error on macOS with system GLEW
The reason why includes are failing is because the GLEW include
directories might not be in the system ones (as it is typically
installed via Homebrew).

The way CMake code was structures it was linking directly to the
GLEW library (instead of using imported target or using the
GLEW_LIBRARIES), and it was not using GLEW_INCLUDE_DIRS.

This change makes it so limeGUI is linked against GLEW::GLEW
imported target which makes it so the GLEW include directories
are configured for the limeGUI. For the case when system-wide GLEW
is not installed an imported target is created from the compiled
GLEW library.

Tested both scenarios (GLEW available and is not available) by
commenting out find_package(GLEW) and doing a full clean build.